+++ src/main/java/com/mcmiddleearth/mcme/pvp/command/CommandNewMapArgument.java
@@ -20,7 +20,7 @@ public class CommandNewMapArgument implements ArgumentType<String> {
     static Set<String> options;
     Set<String> endings;
     public CommandNewMapArgument() {
-        options = PVPCommand.getMapNames(); endings = new HashSet<>(Lists.newArrayList("-TS", "-TDM", "-IN", "-OITQ", "-TC", "-RB", "-FFA", "-DR", "-CTF","-EH","-CQ"));
+        options = PVPCommand.getMapNames(); endings = new HashSet<>(Lists.newArrayList("-TS", "-TDM", "-IN", "-OITQ", "-TC", "-RB", "-FFA", "-DR", "-CTF","-EH","-SG"));
     }
 
     @Override
